face-service/encodings/requirements.txt
flask==3.0.0
face_recognition==1.3.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
gunicorn==21.2.0